requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# Database
sqlalchemy>=2.0.40
aiosqlite>=0.19.0

# SMS
twilio==8.13.0

# Market data
httpx==0.26.0

# AI
anthropic==0.43.0

# Scheduler
apscheduler==3.10.4

# Environment
python-dotenv==1.0.0

# Security
itsdangerous==2.1.2
